backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary

gunicorn

langchain==0.1.16
langchain-community==0.0.32
pgvector
google-genai>=0.3.2
python-dotenv
nest_asyncio
